+++ test/chat_web/live/main_live/index_test.exs
@@ -5,10 +5,12 @@ defmodule ChatWeb.MainLive.IndexTest do
   import Phoenix.LiveViewTest
 
   alias Chat.Admin.MediaSettings
-  alias Chat.{AdminDb, AdminRoom, Card, Db, Dialogs, Rooms, RoomsBroker, User, UsersBroker}
+  alias Chat.{AdminDb, AdminRoom, Card, Db, Dialogs, Rooms, User}
   alias Chat.Dialogs.PrivateMessage
   alias Chat.{Identity, Messages, RoomInviteIndex}
+  alias Chat.Rooms.RoomsBroker
   alias Chat.Sync.{CargoRoom, UsbDriveDumpRoom}
+  alias Chat.User.UsersBroker
 
   describe "room sync between multiple tabs" do
     setup [:current_tab, :another_tab]